# api/openapi.yaml
openapi: 3.1.0
info:
  title: DƏLİL API
  version: 0.1.0
  summary: Cryptographically verifiable audit infrastructure.
  description: |
    Applications send audit events to DƏLİL. Every event is canonicalized (RFC 8785),
    hashed (SHA-256), linked to its predecessor in its stream and signed (Ed25519).
    Verification recomputes all of it and reports exactly where a chain was altered.

    **Authentication.** Machine clients use project API keys
    (`Authorization: Bearer dlk_…`). Dashboard sessions use `Authorization: Bearer dls_…`
    and must name the project with the `Delil-Project` header.

    **Errors** always have the shape `{"error": {"code", "message", "requestId", "details"?}}`.
    Every response carries `X-Request-Id`.

    **Retries.** Event ingestion accepts an `Idempotency-Key`; retrying with the same key
    and body returns the original result with `Idempotent-Replayed: true`.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  contact:
    name: DƏLİL
    url: https://github.com/serxan22/delil
servers:
  - url: http://localhost:8080
    description: Local development (docker compose)
security:
  - bearer: []
...
paths:
...
  /v1/events:
    post:
      tags: [Events]
      summary: Record an event
      description: Requires the `events:write` scope. The stream is created on first use.
      operationId: recordEvent
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/EventInput" } } }
      responses:
        "201":
          description: Committed (or replayed, see `Idempotent-Replayed`).
          headers: { Idempotent-Replayed: { $ref: "#/components/headers/IdempotentReplayed" } }
          content: { application/json: { schema: { $ref: "#/components/schemas/EventReceipt" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/Invalid" }
    get:
      tags: [Events]
      summary: List events (newest first)
      description: Requires `events:read`.
      operationId: listEvents
      parameters:
        - { name: stream, in: query, schema: { type: string } }
        - { name: actorId, in: query, schema: { type: string } }
        - { name: actorType, in: query, schema: { type: string } }
        - { name: action, in: query, description: "Exact action, or a prefix ending with `*`.", schema: { type: string } }
        - { name: resourceType, in: query, schema: { type: string } }
        - { name: resourceId, in: query, schema: { type: string } }
        - { name: from, in: query, description: Recorded at or after (RFC 3339 or YYYY-MM-DD)., schema: { type: string } }
        - { name: to, in: query, description: Recorded before (RFC 3339 or YYYY-MM-DD)., schema: { type: string } }
        - { name: verificationStatus, in: query, schema: { type: string, enum: [verified, failed, unverified] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: A page of events.
          content: { application/json: { schema: { $ref: "#/components/schemas/EventList" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/events/batch:
    post:
      tags: [Events]
      summary: Record several events atomically
      description: |
        Requires `events:write`. All events are committed in one transaction, or none is.
        Events may target different streams.
      operationId: recordBatch
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              additionalProperties: false
              properties:
                events: { type: array, minItems: 1, maxItems: 500, items: { $ref: "#/components/schemas/EventInput" } }
      responses:
        "201":
          description: Committed (or replayed, see `Idempotent-Replayed`).
          headers: { Idempotent-Replayed: { $ref: "#/components/headers/IdempotentReplayed" } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  object: { const: list }
                  data: { type: array, items: { $ref: "#/components/schemas/EventReceipt" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/Invalid" }
  /v1/events/{id}:
    get:
      tags: [Events]
      summary: Get an event
      description: Requires `events:read`.
      operationId: getEvent
      parameters: [{ $ref: "#/components/parameters/EventId" }]
      responses:
        "200": { description: The event, content: { application/json: { schema: { $ref: "#/components/schemas/Event" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/events/{id}/verify:
    get:
      tags: [Verification]
      summary: Verify one event and its links to its neighbours
      description: Requires `verify`. A local check; verify the stream for whole-chain guarantees.
      operationId: verifyEvent
      parameters: [{ $ref: "#/components/parameters/EventId" }]
      responses:
        "200": { description: Report, content: { application/json: { schema: { $ref: "#/components/schemas/VerificationReport" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/streams:
    get:
      tags: [Streams]
      summary: List streams
      description: Requires `events:read`.
      operationId: listStreams
      responses:
        "200": { description: Streams, content: { application/json: { schema: { $ref: "#/components/schemas/StreamList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/streams/{name}:
    get:
      tags: [Streams]
      summary: Get a stream and its head
      description: Requires `events:read`.
      operationId: getStream
      parameters: [{ $ref: "#/components/parameters/StreamName" }]
      responses:
        "200": { description: Stream, content: { application/json: { schema: { $ref: "#/components/schemas/Stream" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/streams/{name}/chain:
    get:
      tags: [Streams]
      summary: Raw chain records for independent verification
      description: |
        Requires `events:read`. Returns headers, hashes, signatures and the exact canonical
        content, in ascending sequence order. The delil CLI uses this endpoint to verify
        without trusting the server.
      operationId: getChain
      parameters:
        - { $ref: "#/components/parameters/StreamName" }
        - { name: afterSequence, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 5000, default: 1000 } }
        - { name: content, in: query, description: "Set to `false` for header-only records.", schema: { type: boolean, default: true } }
      responses:
        "200": { description: A page of records, content: { application/json: { schema: { $ref: "#/components/schemas/ChainPage" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/streams/{name}/verify:
    post:
      tags: [Verification]
      summary: Verify a stream on the server and record the run
      description: Requires `verify`.
      operationId: verifyStream
      parameters: [{ $ref: "#/components/parameters/StreamName" }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                witnesses: { type: array, description: Checkpoints saved earlier; they must still be part of the chain., items: { $ref: "#/components/schemas/Checkpoint" } }
      responses:
        "200": { description: Verification run with its report, content: { application/json: { schema: { $ref: "#/components/schemas/VerificationRun" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/streams/{name}/checkpoints:
    get:
      tags: [Streams]
      summary: List signed checkpoints
      description: Requires `events:read`. Ascending sequence order.
      operationId: listCheckpoints
      parameters: [{ $ref: "#/components/parameters/StreamName" }]
      responses:
        "200": { description: Checkpoints, content: { application/json: { schema: { $ref: "#/components/schemas/CheckpointList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Streams]
      summary: Sign the current head
      description: Requires `verify`. Returns 200 with the existing checkpoint if the head is already checkpointed.
      operationId: createCheckpoint
      parameters: [{ $ref: "#/components/parameters/StreamName" }]
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Checkpoint" } } } }
        "200": { description: Already checkpointed, content: { application/json: { schema: { $ref: "#/components/schemas/Checkpoint" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/verify:
    post:
      tags: [Verification]
      summary: Verify every stream of the project
      description: Requires `verify`.
      operationId: verifyProject
      responses:
        "200": { description: Verification run, content: { application/json: { schema: { $ref: "#/components/schemas/VerificationRun" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/verification-runs:
    get:
      tags: [Verification]
      summary: List verification runs
      description: Requires `events:read`. Newest first; reports are omitted.
      operationId: listVerificationRuns
      parameters: [{ name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }]
      responses:
        "200": { description: Runs, content: { application/json: { schema: { $ref: "#/components/schemas/VerificationRunList" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/verification-runs/{id}:
    get:
      tags: [Verification]
      summary: Get a verification run with its report
      description: Requires `events:read`.
      operationId: getVerificationRun
      parameters: [{ name: id, in: path, required: true, schema: { type: string, pattern: "^vrf_[0-9A-Z]{26}$" } }]
      responses:
        "200": { description: Run, content: { application/json: { schema: { $ref: "#/components/schemas/VerificationRun" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/exports:
    post:
      tags: [Exports]
      summary: Create an evidence package
      description: Requires `exports`. Built asynchronously; poll `GET /v1/exports/{id}`.
      operationId: createExport
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ExportRequest" } } }
      responses:
        "202": { description: Queued, content: { application/json: { schema: { $ref: "#/components/schemas/Export" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "404": { $ref: "#/components/responses/NotFound" }
    get:
      tags: [Exports]
      summary: List exports
      description: Requires `exports`. Newest first.
      operationId: listExports
      responses:
        "200": { description: Exports, content: { application/json: { schema: { $ref: "#/components/schemas/ExportList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/exports/{id}:
    get:
      tags: [Exports]
      summary: Get an export
      description: Requires `exports`.
      operationId: getExport
      parameters: [{ $ref: "#/components/parameters/ExportId" }]
      responses:
        "200": { description: Export, content: { application/json: { schema: { $ref: "#/components/schemas/Export" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/exports/{id}/download:
    get:
      tags: [Exports]
      summary: Download the evidence package (ZIP)
      description: Requires `exports`. 409 while the package is still being built, has failed or has expired.
      operationId: downloadExport
      parameters: [{ $ref: "#/components/parameters/ExportId" }]
      responses:
        "200": { description: The package, content: { application/zip: { schema: { type: string, format: binary } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/api-keys:
    post:
      tags: [Keys]
      summary: Create an API key
      description: |
        Requires `api_keys:manage`. The secret is returned once. An API key can only grant
        scopes it holds itself.
      operationId: createApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              additionalProperties: false
              properties:
                name: { type: string, maxLength: 100 }
                scopes: { type: array, minItems: 1, items: { $ref: "#/components/schemas/Scope" } }
                expiresAt: { type: string, format: date-time }
      responses:
        "201": { description: "Created (includes `secret`, shown once)", content: { application/json: { schema: { $ref: "#/components/schemas/ApiKey" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
    get:
      tags: [Keys]
      summary: List API keys (never includes secrets)
      description: Requires `api_keys:manage`.
      operationId: listApiKeys
      responses:
        "200": { description: Keys, content: { application/json: { schema: { $ref: "#/components/schemas/ApiKeyList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/api-keys/{id}:
    delete:
      tags: [Keys]
      summary: Revoke an API key
      description: Requires `api_keys:manage`.
      operationId: revokeApiKey
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Revoked, content: { application/json: { schema: { $ref: "#/components/schemas/ApiKey" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/signing-keys:
    get:
      tags: [Keys]
      summary: List signing keys
      description: Requires `keys:read`. Oldest first.
      operationId: listSigningKeys
      responses:
        "200": { description: Keys, content: { application/json: { schema: { $ref: "#/components/schemas/SigningKeyList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/signing-keys/export:
    get:
      tags: [Keys]
      summary: Public keys in trusted-keys format (for pinning)
      description: Requires `keys:read`. Save the result and pass it to `delil verify --trusted-keys`.
      operationId: exportSigningKeys
      responses:
        "200": { description: Keys file, content: { application/json: { schema: { $ref: "#/components/schemas/KeysFile" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/signing-keys/rotate:
    post:
      tags: [Keys]
      summary: Rotate the project's signing key
      description: Requires `keys:rotate`. The old key is retired but kept, so history stays verifiable.
      operationId: rotateSigningKey
      responses:
        "200":
          description: Rotation result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  previous: { $ref: "#/components/schemas/SigningKey" }
                  current: { $ref: "#/components/schemas/SigningKey" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/signing-keys/{id}/revoke:
    post:
      tags: [Keys]
      summary: Revoke a compromised signing key
      description: |
        Requires `keys:rotate`. Events recorded after the revocation time with this key fail
        verification. Revoking the active key creates a replacement.
      operationId: revokeSigningKey
      parameters: [{ name: id, in: path, required: true, schema: { type: string, pattern: "^ed25519:[0-9a-f]{32}$" } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [reason], properties: { reason: { type: string, minLength: 1, maxLength: 500 } } }
      responses:
        "200": { description: Revoked, content: { application/json: { schema: { $ref: "#/components/schemas/KeyRevocation" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/project:
    get:
      tags: [Projects]
      summary: The current project and its settings
      description: Available to any API key or session of the project.
      operationId: getProject
      responses:
        "200": { description: Project, content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
    patch:
      tags: [Projects]
      summary: Update the project name or ingestion settings (dashboard admins)
      description: Requires `project:manage`, which only dashboard administrators hold.
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: { type: string }
                settings: { $ref: "#/components/schemas/ProjectSettings" }
      responses:
        "200": { description: Updated project, content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/overview:
    get:
      tags: [Projects]
      summary: Statistics and integrity health
      description: Requires `events:read`.
      operationId: getOverview
      responses:
        "200": { description: Overview, content: { application/json: { schema: { type: object } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /v1/auth/login:
    post:
      tags: [Sessions]
      summary: Exchange email and password for a session token
      description: Rate limited per client IP and per email. Unknown accounts and wrong passwords return the same error.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [email, password], properties: { email: { type: string }, password: { type: string } } }
      responses:
        "200": { description: "Session, including the token (shown once)", content: { application/json: { schema: { $ref: "#/components/schemas/Session" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: A project API key (`dlk_…`) or a dashboard session token (`dls_…`).
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 1-255 printable ASCII characters. Retries with the same key and body return the original result.
      schema: { type: string, maxLength: 255 }
...
  headers:
    IdempotentReplayed:
      description: "`true` when the response replays an earlier request with the same idempotency key."
      schema: { type: string }
  responses:
...
    Forbidden: { description: Missing scope, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...